Quality Alloys Group Presentation
In our Data Science: R course at Hult, we had a group presentation on the business case 'Web Analytics at Quality Alloys Inc.'
Together we utilized what we had learned in data cleaning, transformations, analysis, and visualizations.

The deliverables were to be responses to the questions in the case, an analysis of the effectiveness of the business' website, and recommendations on how to improve sales. 

The spreadsheet, business case, final submission summary, and ppt presentation are in this repository.